chore: update playwright
