+++ client/package.json
@@ -12,14 +12,14 @@
     "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
     "test": "vitest run",
     "start-server": "cd .. && make serve",
-    "test-e2e": "playwright test -c ./src/playwright.config.ts",
+    "test-e2e": "playwright test",
     "test-e2e:ci": "start-server-and-test start-server '3579/docs|3000' test-e2e",
     "test:watch": "vitest watch",
     "test:coverage": "npm run test -- --coverage"
   },
   "devDependencies": {
     "@fullhuman/postcss-purgecss": "^4.1.3",
-    "@playwright/test": "^1.25.1",
+    "@playwright/test": "^1.27.1",
     "@sveltejs/adapter-node": "^1.0.0-next.90",
     "@sveltejs/kit": "^1.0.0-next.481",
     "@testing-library/jest-dom": "^5.16.5",
