+++ client/playwright.config.ts
@@ -1,0 +1,54 @@
+import path from "path";
+import dotenv from "dotenv";
+import type { PlaywrightTestConfig } from "@playwright/test";
+import type { AppTestArgs } from "./src/tests/e2e/fixtures";
+
+dotenv.config({
+  path: path.resolve("..", ".env"),
+});
+
+const config: PlaywrightTestConfig<AppTestArgs> = {
+  testDir: "src/tests/e2e/",
+  globalSetup: "src/tests/e2e/global-setup.ts",
+  retries: 3,
+  expect: {
+    /**
+     * Maximum time expect() should wait for the condition to be met.
+     * For example in `await expect(locator).toHaveText();`
+     */
+    timeout: 5000
+  },
+  /* Run tests in files in parallel */
+  fullyParallel: true,
+  /* Fail the build on CI if you accidentally left test.only in the source code. */
+  forbidOnly: !!process.env.CI,
+  /* Opt out of parallel tests on CI. */
+  workers: process.env.CI ? 1 : undefined,
+  /* Reporter to use. See https://playwright.dev/docs/test-reporters */
+  reporter: 'html',
+  /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
+  use: {
+    baseURL: "http://localhost:3000",
+    browserName: "firefox",
+    headless: true,
+    screenshot: "only-on-failure",
+    video: "on-first-retry",
+    /* Maximum time each action such as `click()` can take. Defaults to 0 (no limit). */
+    actionTimeout: 0,
+    /* Base URL to use in actions like `await page.goto('/')`. */
+    // baseURL: 'http://localhost:3000',
+
+    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
+    trace: 'on-first-retry',
+  },
+  /* Folder for test artifacts such as screenshots, videos, traces, etc. */
+  // outputDir: 'test-results/',
+
+  /* Run your local dev server before starting the tests */
+  // webServer: {
+  //   command: 'npm run start',
+  //   port: 3000,
+  // },
+};
+
+export default config;
